Time measurements added to find out, why the permanent polling doesn't generate high CPU load.
Finding: Without CAN Rx, the code takes about 3.38 us per polling cycle and spends 3.23 us inside of PCANBasic.Read. If we don't have an effect of the CPU load measurement in Task Manager then it seems that the call of Read has an implicit none-busy wait for the result of about 3 us. This short release of the CPU may suffice to avoid blocking the entire CPU.