Dead end: Attempt to use SetValue(PCAN_RECEIVE_EVENT) to implement busy-wait free fast responsiveness on incoming DTO messages.
Unfortunately, the Java wrapper of the PCAN DLL seems not to support the mechanism. SetValue returns "PCAN_ERROR_ILLPARAMTYPE: An invalid parameter was specified". A Java sample indicates that event based processing is only possible using a callback mechanism, which triggers a callback on Rx in another Java thread. This would unfortunately require a totally different code structure. The code in this revision could still help, as it deals with Windows events. A possibility to rescue the current code architecture is using the callback just to notify a Windows event and waiting for the event in our otherwise single-threaded application. With other words, building the mechanism ourselves, which has been promised by the PCAN documentation for parameter PCAN_RECEIVE_EVENT.