Application interface (CLI) elaborated.
The structure of the CLI is now basically done. We have three modes, enumeration of CAN devices, up- and download. The use of certain arguments selects the mode, e.g., an srec input file commands the download and flash programming. Upload is now controlled via direct address range specification on the command line.
  Tool version becomes 0.8